requirements.txt
streamlit
torch
torchvision
transformers
Pillow
openai
